Revert "feat: install gcal + gvideo by default for google signups (#17773)"
This reverts commit fcd06c80e3998762d973266325b6fd2ed352e6a1.
