+++ .env.example
@@ -119,7 +119,7 @@ NEXT_PUBLIC_FRESHCHAT_HOST=
 
 # Google OAuth credentials
 # To enable Login with Google you need to:
-# 1. Set `GOOGLE_API_CREDENTIALS` below
+# 1. Set `GOOGLE_API_CREDENTIALS` above
 # 2. Set `GOOGLE_LOGIN_ENABLED` to `true`
 # When self-hosting please ensure you configure the Google integration as an Internal app so no one else can login to your instance
 # @see https://support.google.com/cloud/answer/6158849#public-and-internal&zippy=%2Cpublic-and-internal-applications
