+++ packages/app-store/googlecalendar/lib/CalendarService.ts
@@ -18,9 +18,10 @@ import {
   CREDENTIAL_SYNC_SECRET_HEADER_NAME,
 } from "@calcom/lib/constants";
 import { formatCalEvent } from "@calcom/lib/formatCalendarEvent";
+import { getAllCalendars } from "@calcom/lib/google";
 import logger from "@calcom/lib/logger";
 import { safeStringify } from "@calcom/lib/safeStringify";
-import { GoogleService } from "@calcom/lib/server/service/google";
+import { GoogleRepository } from "@calcom/lib/server/repository/google";
 import prisma from "@calcom/prisma";
 import type {
   Calendar,
@@ -542,7 +543,7 @@ export default class GoogleCalendarService implements Calendar {
     }
     async function getCalIds() {
       if (selectedCalendarIds.length !== 0) return selectedCalendarIds;
-      const cals = await GoogleService.getAllCalendars(calendar, ["id"]);
+      const cals = await getAllCalendars(calendar, ["id"]);
       if (!cals.length) return [];
       return cals.reduce((c, cal) => (cal.id ? [...c, cal.id] : c), [] as string[]);
     }
@@ -606,7 +607,7 @@ export default class GoogleCalendarService implements Calendar {
             status: 200,
             statusText: "OK",
             data: {
-              items: await GoogleService.getAllCalendars(calendar),
+              items: await getAllCalendars(calendar),
             },
           })
       );
@@ -651,7 +652,7 @@ export default class GoogleCalendarService implements Calendar {
       },
     });
     const response = res.data;
-    await GoogleService.upsertSelectedCalendar({
+    await GoogleRepository.upsertSelectedCalendar({
       userId: this.credential.userId!,
       externalId: calendarId,
       credentialId: this.credential.id,
@@ -685,7 +686,7 @@ export default class GoogleCalendarService implements Calendar {
       .catch((err) => {
         console.warn(JSON.stringify(err));
       });
-    await GoogleService.upsertSelectedCalendar({
+    await GoogleRepository.upsertSelectedCalendar({
       userId: this.credential.userId!,
       externalId: calendarId,
       credentialId: this.credential.id,
