+++ packages/lib/server/repository/credential.ts
@@ -30,15 +30,6 @@ export class CredentialRepository {
     });
   }
 
-  static async findFirstByAppIdAndUserId({ appId, userId }: { appId: string; userId: number }) {
-    return await prisma.credential.findFirst({
-      where: {
-        appId,
-        userId,
-      },
-    });
-  }
-
   static async findFirstByUserIdAndType({ userId, type }: { userId: number; type: string }) {
     return await prisma.credential.findFirst({ where: { userId, type } });
   }
