+++ packages/lib/server/repository/selectedCalendar.ts
@@ -3,8 +3,15 @@ import type { Prisma } from "@prisma/client";
 import { prisma } from "@calcom/prisma";
 import { credentialForCalendarServiceSelect } from "@calcom/prisma/selects/credential";
 
+type SelectedCalendarCreateInput = {
+  credentialId: number;
+  userId: number;
+  externalId: string;
+  integration: string;
+};
+
 export class SelectedCalendarRepository {
-  static async create(data: Prisma.SelectedCalendarUncheckedCreateInput) {
+  static async create(data: SelectedCalendarCreateInput) {
     return await prisma.selectedCalendar.create({
       data: {
         ...data,
